Fix compiler warning (signed/unsigned comparison)
This commit was SVN r30142.

+++ ompi/mca/btl/vader/btl_vader_module.c
@@ -313,7 +313,7 @@ static int vader_del_procs(struct mca_btl_base_module_t *btl,
                            size_t nprocs, struct ompi_proc_t **procs,
                            struct mca_btl_base_endpoint_t **peers)
 {
-    for (int i = 0 ; i < nprocs ; ++i) {
+    for (size_t i = 0 ; i < nprocs ; ++i) {
         fini_vader_endpoint (peers[i]);
         peers[i] = NULL;
     }
